spliting the data layer
